batch processor and related classes moved out of audit project
