+++ src/SFA.DAS.Payments.Audit.Application.ServiceFabric/PaymentsEventModelCache/PaymentsEventModelCache.cs
@@ -4,11 +4,13 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.ServiceFabric.Data.Collections;
+using SFA.DAS.Payments.Application.BatchWriting;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
 using SFA.DAS.Payments.Audit.Application.PaymentsEventModelCache;
 using SFA.DAS.Payments.Audit.Application.ServiceFabric.Infrastructure;
 using SFA.DAS.Payments.Audit.Model;
 using SFA.DAS.Payments.ServiceFabric.Core;
+using SFA.DAS.Payments.ServiceFabric.Core.BatchWriting;
 
 namespace SFA.DAS.Payments.Audit.Application.ServiceFabric.PaymentsEventModelCache
 {
