+++ src/SFA.DAS.Payments.Audit.Application/PaymentsEventProcessing/PaymentsEventModelBatchService.cs
@@ -5,6 +5,7 @@ using Polly;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
 using SFA.DAS.Payments.Audit.Model;
 using SFA.DAS.Payments.Core.Configuration;
+using SFA.DAS.Payments.ServiceFabric.Core.BatchWriting;
 
 namespace SFA.DAS.Payments.Audit.Application.PaymentsEventProcessing
 {
