+++ src/SFA.DAS.Payments.DataLocks.DataLockEventService/DataLockEventService.cs
@@ -1,22 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Fabric;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Autofac;
 using Microsoft.ServiceFabric.Actors.Runtime;
-using Microsoft.ServiceFabric.Data.Collections;
 using Microsoft.ServiceFabric.Services.Communication.Runtime;
 using Microsoft.ServiceFabric.Services.Runtime;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
-using SFA.DAS.Payments.Audit.Application.ServiceFabric.Infrastructure;
 using SFA.DAS.Payments.ServiceFabric.Core;
 
 namespace SFA.DAS.Payments.DataLocks.DataLockEventService
 {
     [StatePersistence(StatePersistence.Persisted)]
-    internal sealed class DataLockEventService : AuditStatefulService<DataLockStatusChanged>
+    internal sealed class DataLockEventService : StatefulService
     {
         private IStatefulEndpointCommunicationListener listener;
         private readonly ILifetimeScope lifetimeScope;
