+++ src/SFA.DAS.Payments.EarningEvents.Application/Infrastructure/Ioc/EarningEventsModule.cs
@@ -1,4 +1,5 @@
 ﻿using Autofac;
+using SFA.DAS.Payments.Application.BatchWriting;
 using SFA.DAS.Payments.Application.Data.Configurations;
 using SFA.DAS.Payments.Application.Repositories;
 using SFA.DAS.Payments.EarningEvents.Application.Interfaces;
