+++ src/SFA.DAS.Payments.ServiceFabric.Core/BatchWriting/BatchScope.cs
@@ -1,16 +1,11 @@
 ﻿using System;
 using System.Threading.Tasks;
-using System.Transactions;
 using Autofac;
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.ApplicationInsights.Extensibility;
 using SFA.DAS.Payments.Application.Infrastructure.Telemetry;
-using SFA.DAS.Payments.Audit.Application.PaymentsEventProcessing;
-using SFA.DAS.Payments.Audit.Application.ServiceFabric.Infrastructure;
-using SFA.DAS.Payments.Audit.Model;
-using SFA.DAS.Payments.ServiceFabric.Core;
 
-namespace SFA.DAS.Payments.Audit.Application.ServiceFabric.PaymentsEventProcessing
+namespace SFA.DAS.Payments.ServiceFabric.Core.BatchWriting
 {
     public class BatchScope: IBatchScope
     {
@@ -37,9 +32,9 @@ namespace SFA.DAS.Payments.Audit.Application.ServiceFabric.PaymentsEventProcessi
             lifetimeScope?.Dispose();
         }
 
-        public IPaymentsEventModelBatchProcessor<T> GetBatchProcessor<T>() where T: IPaymentsEventModel
+        public IBatchProcessor<T> GetBatchProcessor<T>()
         {
-            return lifetimeScope.Resolve<IPaymentsEventModelBatchProcessor<T>>();
+            return lifetimeScope.Resolve<IBatchProcessor<T>>();
         }
 
         public void Abort()
