+++ src/SFA.DAS.Payments.ServiceFabric.Core/BatchWriting/IBatchProcessor.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Payments.ServiceFabric.Core.BatchWriting
+{
+    public interface IBatchProcessor<T>
+    {
+        Task<int> Process(int batchSize, CancellationToken cancellationToken);
+    }
+}
