+++ src/SFA.DAS.Payments.ServiceFabric.Core/BatchWriting/IBatchScope.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Payments.ServiceFabric.Core.BatchWriting
+{
+    public interface IBatchScope: IDisposable
+    {
+        IBatchProcessor<T> GetBatchProcessor<T>();
+        void Abort();
+        Task Commit();
+    }
+}
