+++ src/SFA.DAS.Payments.ServiceFabric.Core/BatchWriting/StateManagerUnitOfWork.cs
@@ -3,9 +3,8 @@ using System.Threading.Tasks;
 using Microsoft.ServiceFabric.Data;
 using NServiceBus.UnitOfWork;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
-using SFA.DAS.Payments.ServiceFabric.Core;
 
-namespace SFA.DAS.Payments.Audit.Application.ServiceFabric.Infrastructure
+namespace SFA.DAS.Payments.ServiceFabric.Core.BatchWriting
 {
     public class StateManagerUnitOfWork : IManageUnitsOfWork
     {
